requirements.txt
pyyaml
requests
feedparser
sqlalchemy
mysqlclient
flask

